backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.22
botocore==1.35.22
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
outcome==1.3.0.post0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.2
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
